+++ internals-js/src/precompute.ts
@@ -5,8 +5,6 @@ import {
   federationMetadata,
   FieldDefinition,
   collectTargetFields,
-  InterfaceType,
-  ObjectType,
   Schema,
 } from ".";
 
@@ -33,7 +31,7 @@ export function computeShareables(schema: Schema): (field: FieldDefinition<Compo
     }
   };
 
-  for (const type of schema.types<ObjectType>('ObjectType')) {
+  for (const type of schema.objectTypes()) {
     addKeyFields(type);
     const shareablesOnType = shareableDirective ? type.appliedDirectivesOf(shareableDirective) : [];
     for (const field of type.fields()) {
@@ -59,7 +57,7 @@ export function computeShareables(schema: Schema): (field: FieldDefinition<Compo
     }
   }
 
-  for (const type of schema.types<InterfaceType>('InterfaceType')) {
+  for (const type of schema.interfaceTypes()) {
     addKeyFields(type);
   }