+++ query-graphs-js/src/index.ts
@@ -2,7 +2,6 @@ export * from './querygraph';
 export * from './graphPath';
 export * from './pathTree';
 export * from './graphviz';
-export * from './structuralSubtyping';
 export * from './transition';
 export * from './pathContext';
 export * from './conditionsCaching';